Add DNS records for environment
Forward to API Gateway and use the DNS endpoint for tests.

I've weighted 100% to eu-west-1 for now because multi-region signatures aren't well supported yet (i.e. you need to know which region you're signing for) and all our traffic originates in eu-west-1, but we should really take a proper look at this for future reliability.

#minor