Rewrite Go modules and Dockerfiles
`go.work` has a bunch of annoying side effects on building, namely that it pushes you to include everything in the workspace in your build images.

Instead, promote all the dependencies to a single `go.mod` file in the top level, and then build an individual directory in the Dockerfiles.

This simplifies the dependency config and allows us to build images with a smaller subset of files so that we can rebuild more quickly.

#minor