Add visual editor to fixtures
Use the JSON schema to generate a bunch of nested fields so that users don't need to understand JSON in order to create fixture data in the store.

For now I've hard-coded the current JSON schema for ease, but I'm hoping to work on formalising all the schemas properly soon.

Fixes CTC-145 #minor