Split out LPA definition into JSON Schema
Create a new schemas directory and move the core LPA definition there. Import it into the OpenAPI schema and also import it into the fixtures container.

This is part of a longer journey to build well-defined reusable schemas that are stored alongside an LPA and can be used for validation.

For CTC-145 #minor