Add separate OpenAPI file for API Gateway
APIGW only supports OpenAPI 3.0, but we need to use 3.1 to import an external JSON Schema document.

To allow us to still build APIGW from OpenAPI, we need to have two files: a proper one that we use for validation/mocking and one for AWS.

We can't extend or merge OpenAPI schemas, which would have been a nice solution where the AWS one would have defined the basics and the main one could add request/response schemas.

#minor