Split out LPA definition into JSON Schema (#157)
* Split out LPA definition into JSON Schema

Create a new schemas directory and move the core LPA definition there. Import it into the OpenAPI schema and also import it into the fixtures container.

This is part of a longer journey to build well-defined reusable schemas that are stored alongside an LPA and can be used for validation.

For CTC-145 #minor

* Add separate OpenAPI file for API Gateway

APIGW only supports OpenAPI 3.0, but we need to use 3.1 to import an external JSON Schema document.

To allow us to still build APIGW from OpenAPI, we need to have two files: a proper one that we use for validation/mocking and one for AWS.

We can't extend or merge OpenAPI schemas, which would have been a nice solution where the AWS one would have defined the basics and the main one could add request/response schemas.

#minor

* Resolve conflicts with main

#patch

* Compile APIGW OpenAPI spec in pipeline

Rather than manually maintaining a separate OpenAPI file, build it in th pipeline by merging the base file with overrides to support v3.0.3.

Because we're merging the specs, we can't remove the external `$ref`. So instead replace it with a local pointer to an empty object.

#minor