Fix default values in fixtures app (#250)
Currently the values shown on screen are auto-filled based on the schema. This includes selecting the first value of enum dropdowns. But the data saved in the textarea (which is ultimately sent to the API) doesn't have these which can cause confusion.

To ensure alignment, after filling in default values save the new data back to the textarea so that it represents what's actually being displayed.

This makes the textarea slightly harder to edit manually, but no-one's doing that anyway.

This replaces the `requestAnimationFrame` fix which just caused different issues.

#patch