Default date fields to null
json-schema-library is defaulting al text fields to empty strings. This is fine for general text fields, but causes problems for dates as the the API then tries to unmarshal the string and finds it isn't a valid ISO-8601 date.

Empty date strings should instead be treated as `null` so that unmarshaling is skipped.

Fixes VEGA-2708 #patch