Remove target event bus from default env config (#284)
The "default" environment configuration is used by pop-up environments. It should not have a target event bus because there's no valid bus to send to.

The existing config (sending to the Sirius dev event bus) is invalid because Sirius can't then read the LPA back from the store: it tries—and fails—to use the dev LPA Store, not the pop-up.

Fixes VEGA-2790 #patch