Ensure actors have UIDs on creation (#336)
* Ensure actors have UIDs on creation

If any actors have not been assigned a UID when the case is created, automatically add one.

This ensures they will have a consistent identity for all future events and interactions.

Fixes VEGA-2988 #minor

* Generate UIDs for remaining actors

Also assigned UIDs to people to notify and the independent witness and authorised signatory if they have been provided.

#patch